generic irq improvements
This adds the possibility to specify in advance an event that is tied
to a specific user-provided callback. This callback is simplified and
now only takes a void *ctx parameter.

If registering a callback is done with the callback descriptor, then
it makes sense to also unregister a callback by providing the callback
descriptor.

Signed-off-by: Darius Berghe <darius.berghe@analog.com>