drivers:platform:maxim:maxim_uart.c: UART block
This change makes the no_os_uart_write to finish when
all the bytes were transmitted. Before this, the function
only blocked until all the bytes were in the TX fifo, which
caused a problem when using a read_nonblocking after the
no_os_uart_write, since that flushes both TX and RX fifos.

Signed-off-by: Ciprian Regus <ciprian.regus@analog.com>