drivers:platform:maxim: Replaced fifo read/write
Replaced the FIFO reading/writing in a while loop with the MXC_UART_Read
and MXC_UART_Write functions. This implementation doesn't allow for
returning the actual number of bytes read/written, but the FIFO access
is faster.

With the previous implementation, the communication with an iio client
would time out after a while, when continuously trasmitting data.

Signed-off-by: Ciprian Regus <ciprian.regus@analog.com>