add PID control utility to no-OS
This PID supports:
- setting Kp, Ki, Kd gains
- clipping output to a certain range to match the range of whatever PID
  is controlling (for example, for an 8-bit PWM, the algorithm would
  clip the output to a range between 0 and 255).
- hysteresis
- clipping internal integrator to a certain range to avoid overflow

The no_os_pid_control() must be caled in a loop and user is responsible
for detecting when the control should stop by not calling it anymore.

This can be done either by characterizing the algorithm for a particular
set of configuration parameters and finding a minimum number of
iterations that make the output stable or by looking at the output and
detecting when it sits within a satisfactory tolerance from the set
point.

Signed-off-by: Darius Berghe <darius.berghe@analog.com>