drivers:platform:maxim: Fix UART UB
Currently, when setting the voltage level of the UART interface, it is
done by changing the vssel field of a struct defined in the SDK (which
is used to configure some GPIOs as UART alternate function). However,
these structs are defined as const, and as such, modifying it's fields
is undefined behavior.

The initialization sequence is still correct, since the voltage level is
set after the UART interface is configured.

Fixes: f724568 ("platform: maxim: Remove discarded const warning")
Signed-off-by: Ciprian Regus <ciprian.regus@analog.com>